Update hilt dependencies
