--- app/src/androidTest/java/com/studhub/app/CustomTestRunner.kt
@@ -1,14 +1,0 @@
-package com.studhub.app//import android.app.Application
-//import android.content.Context
-//import androidx.test.runner.AndroidJUnitRunner
-//import dagger.hilt.android.testing.HiltTestApplication
-//
-//class CustomTestRunner : AndroidJUnitRunner() {
-//    override fun newApplication(
-//        cl: ClassLoader?,
-//        className: String?,
-//        context: Context?
-//    ): Application {
-//        return super.newApplication(cl, HiltTestApplication::class.java.name, context)
-//    }
-//}
