+++ build.gradle
@@ -3,7 +3,7 @@ buildscript {
         compose_ui_version = '1.3.3'
         m3_version = '1.0.1'
         m2_version = '1.3.1'
-        hilt_version = '2.44.2'
+        hilt_version = '2.45'
         hilt_navigation_compose_version = '1.0.0'
     }
     repositories {
